Code refactor and add cart button.
