+++ app/src/main/java/com/example/candyshop/data/ContentItem.kt
@@ -1,0 +1,9 @@
+package com.example.candyshop.data
+
+data class ContentItem(
+    val id: Int,
+    val name: String,
+    val image: Int,
+    val price: Int,
+    val description: String = "some placeholder text"
+)
