+++ app/src/main/java/com/example/candyshop/data/contentItems.kt
@@ -1,12 +1,7 @@
-package com.example.candyshop
+package com.example.candyshop.data
 
-data class ContentItem(
-    val id: Int,
-    val name: String,
-    val image: Int,
-    val price: Int,
-    val description: String = "some placeholder text"
-)
+import com.example.candyshop.R
+
 fun getCandyDetailsById(id: Int?): ContentItem? = content.firstOrNull { id == it.id }
 
 val content = listOf(
